requirements.txt
# Core
requests>=2.31.0
beautifulsoup4>=4.12.0
selenium>=4.15.0
lxml>=5.1.0

# LLM
ollama>=0.4.0

# Database
sqlalchemy>=2.0.0

# Scheduling
apscheduler>=3.10.0

# Dashboard
streamlit>=1.30.0

# API
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
pydantic>=2.10.0

# Dedup
rapidfuzz>=3.6.0

# Excel export
openpyxl>=3.1.0

# Email discovery
# apollo-api  # No official package - uses requests
# hunter-api  # No official package - uses requests

# Testing
pytest>=8.0.0
pytest-cov>=4.1.0

# LinkedIn (optional)
# linkedin-api>=2.0.0

# Utilities
python-dotenv>=1.0.0
